requirements.txt
#logic
datetime
pandas
numpy
seaborn
requests
google-cloud-bigquery
google-cloud-storage
db-dtypes
plotly
prophet
dash
#dash_html_components==2.0.0
#dash_core_components==2.0.0


# API
#fastapi
#pytz
#uvicorn

#deploy
#mlflow==2.1.1
#python-dotenv
prefect==2.8.1
#psycopg2-binary


#frontend
streamlit
Pillow
